+++ tests/testthat/test_plots_summary_twa.R
@@ -79,8 +79,8 @@ test_that("Plotting mmkin objects is reproducible", {
 context("AIC calculation")
 
 test_that("The AIC is reproducible", {
-  expect_equivalent(AIC(fits[["SFO", "FOCUS_C"]]), 59.8, scale = 1, tolerance = 0.1)
+  expect_equivalent(AIC(fits[["SFO", "FOCUS_C"]]), 59.3, scale = 1, tolerance = 0.1)
   expect_equivalent(AIC(fits[, "FOCUS_C"]), 
-                    data.frame(df = c(3, 4, 5, 5), AIC = c(59.8, 44.7, 29.1, 39.3)), 
+                    data.frame(df = c(3, 4, 5, 5), AIC = c(59.3, 44.7, 29.0, 39.2)), 
                     scale = 1, tolerance = 0.1)
 })